# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Inferna Next
  version: 0.1.0
paths:
...
  /api/v1/workers:
    get:
      tags:
      - workers
      summary: List Workers
      operationId: list_workers_api_v1_workers_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: cluster_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Cluster Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkerOut'
                title: Response List Workers Api V1 Workers Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/workers/{worker_id}:
    delete:
      tags:
      - workers
      summary: Delete Worker
      operationId: delete_worker_api_v1_workers__worker_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: worker_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Worker Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GPUOut:
      properties:
        id:
          type: integer
          title: Id
        index:
          type: integer
          title: Index
        vendor:
          type: string
          title: Vendor
        name:
          type: string
          title: Name
        vram_mb:
          type: integer
          title: Vram Mb
        used_vram_mb:
          type: integer
          title: Used Vram Mb
        utilization_pct:
          type: integer
          title: Utilization Pct
        uuid:
          anyOf:
          - type: string
          - type: 'null'
          title: Uuid
        driver_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Driver Version
      type: object
      required:
      - id
      - index
      - vendor
      - name
      - vram_mb
      - used_vram_mb
      - utilization_pct
      - uuid
      - driver_version
      title: GPUOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InstanceOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        model_id:
          type: string
          format: uuid
          title: Model Id
        cluster_id:
          type: string
          format: uuid
          title: Cluster Id
        worker_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Worker Id
        engine:
          type: string
          title: Engine
        profile:
          type: string
          title: Profile
        gpu_indexes:
          items:
            type: integer
          type: array
          title: Gpu Indexes
        state:
          type: string
          title: State
        port:
          anyOf:
          - type: integer
          - type: 'null'
          title: Port
        error_detail:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Detail
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        model:
          anyOf:
          - $ref: '#/components/schemas/ModelOut'
          - type: 'null'
        worker_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Worker Name
      type: object
      required:
      - id
      - model_id
      - cluster_id
      - worker_id
      - engine
      - profile
      - gpu_indexes
      - state
      - port
      - error_detail
      - created_at
      - updated_at
      title: InstanceOut
...
    ModelOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        display_name:
          type: string
          title: Display Name
        category:
          type: string
          title: Category
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        params_b:
          anyOf:
          - type: number
          - type: 'null'
          title: Params B
        vram_required_mb:
          type: integer
          title: Vram Required Mb
        requires_hf_token:
          type: boolean
          title: Requires Hf Token
        license:
          anyOf:
          - type: string
          - type: 'null'
          title: License
        is_builtin:
          type: boolean
          title: Is Builtin
      type: object
      required:
      - id
      - name
      - display_name
      - category
      - description
      - params_b
      - vram_required_mb
      - requires_hf_token
      - license
      - is_builtin
      title: ModelOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkerOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        cluster_id:
          type: string
          format: uuid
          title: Cluster Id
        name:
          type: string
          title: Name
        hostname:
          type: string
          title: Hostname
        state:
          type: string
          title: State
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
        os:
          anyOf:
          - type: string
          - type: 'null'
          title: Os
        cpu_cores:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cpu Cores
        memory_mb:
          anyOf:
          - type: integer
          - type: 'null'
          title: Memory Mb
        last_seen_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Seen At
        gpus:
          items:
            $ref: '#/components/schemas/GPUOut'
          type: array
          title: Gpus
          default: []
        instances:
          items:
            $ref: '#/components/schemas/InstanceOut'
          type: array
          title: Instances
          default: []
      type: object
      required:
      - id
      - cluster_id
      - name
      - hostname
      - state
      - version
      - os
      - cpu_cores
      - memory_mb
      - last_seen_at
      title: WorkerOut
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login